+++ src/security/RandomDevice.cpp
@@ -15,10 +15,8 @@
 namespace qindesign {
 namespace security {
 
-RandomDevice &RandomDevice::instance() {
-  static RandomDevice instance;
-  return instance;
-}
+// The UniformRandomBitGenerator instance.
+STATIC_INIT_DEFN(RandomDevice, randomDevice);
 
 FLASHMEM RandomDevice::RandomDevice() {
   if (!trng_is_started()) {
@@ -42,11 +40,6 @@ RandomDevice::result_type RandomDevice::operator()() {
 
 namespace qindesign {
 namespace security {
-
-RandomDevice &RandomDevice::instance() {
-  static RandomDevice instance;
-  return instance;
-}
 
 FLASHMEM RandomDevice::RandomDevice() {
 #if defined(__IMXRT1062__)